code/webvpnbot/requirements.txt
flask
python-telegram-bot
sqlalchemy
python-dotenv
cryptography
